+++ lib/client.ts
@@ -29,7 +29,7 @@ export default class Client {
     usernamePart?: string;
     passwordPart?: string;
     users: any;
-    events: any;
+    events: Event;
     companies: any;
     contacts: Contact;
     leads: any;